The Flop counts for ewald are now included in the total. This was tricky for the CUDA version of Ewald.

+++ ewald.h
@@ -3,7 +3,7 @@
 
 #include "pkd.h"
 
-int pkdParticleEwald(PKD pkd, double *r, float *pa, float *pPot);
+double pkdParticleEwald(PKD pkd, double *r, float *pa, float *pPot);
 void pkdEwaldInit(PKD pkd,int nReps,double fEwCut,double fhCut);
 
 #endif
